Split relational module from core
