+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod core;
 pub mod error;
 pub mod executor;
 pub mod planner;
+pub(crate) mod relational;
 pub mod session;
 pub mod sql_parser;
 pub(crate) mod storage;
