+++ src/relational/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod catalog;
+pub(crate) mod catalog_manager;
+pub(crate) mod cursor;
+pub(crate) mod index_manager;
+pub(crate) mod record_manager;
+pub(crate) mod tuple;
